gcc warning: src/rtlib/unix/thread_core.c
- warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  stack_size = stack_size >= FBTHREAD_STACK_MIN ? stack_size : FBTHREAD_STACK_MIN

+++ src/rtlib/unix/thread_core.c
@@ -58,7 +58,7 @@ FBCALL FBTHREAD *fb_ThreadCreate( FB_THREADPROC proc, void *param, ssize_t stack
 	}
 
 	/* see fb_private_thread.h for defintion of FBTHREAD_STACK_MIN */
-	stack_size = stack_size >= FBTHREAD_STACK_MIN ? stack_size : FBTHREAD_STACK_MIN;
+	stack_size = stack_size >= (ssize_t)FBTHREAD_STACK_MIN ? stack_size : (ssize_t)FBTHREAD_STACK_MIN;
 
 	pthread_attr_setstacksize( &tattr, stack_size );
